+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 -----------
 
 ### Internals
+* Changed the metrics timers to more precisely report in nanoseconds, instead of seconds. ([#3359](https://github.com/realm/realm-core/issues/3359))
 * Better performance when cloud query metrics are turned on, by not acquiring a backtrace on query serialization errors (permissions queries). ([#3361](https://github.com/realm/realm-core/issues/3361)).
 
 ----------------------------------------------